refactor: コンポーネント階層を今のやり方に変更(#8)

+++ src/components/model/Todo/TodoList/TodoList.tsx
@@ -1,6 +1,6 @@
 import { VFC, useCallback } from 'react';
 import { css } from '@emotion/react';
-import TodoListItem from '@/components/TodoList/TodoListItem';
+import TodoListItem from '@/components/model/Todo/TodoListItem';
 import { TodoStatus, Todo } from '@/models/Todo';
 
 type Props = {
@@ -27,7 +27,9 @@ const TodoLlist: VFC<Props> = ({
 
   const handleDeleteTodo = useCallback(
     (id: string) => {
-      deleteTodoFunc(id);
+      if (deleteTodoFunc) {
+        deleteTodoFunc(id);
+      }
     },
     [deleteTodoFunc]
   );
